backend/requirements.txt
# Core Django framework
Django==4.2.17

# Django REST Framework — provides the tools to build RESTful APIs
# (serializers, viewsets, routers, response classes)
djangorestframework==3.15.2

# Adds CORS headers to responses so the frontend (different port/origin)
# can make API requests to the backend without being blocked by the browser
django-cors-headers==4.6.0

# PostgreSQL database adapter for Python
# Using the 'binary' variant bundles its own libpq — no system-level
# PostgreSQL client install needed inside the Docker image
psycopg2-binary==2.9.10

# Production WSGI server — runs multiple worker processes to handle concurrent
# requests. Django's built-in runserver is single-threaded and not for production.
gunicorn==21.2.0
